+++ js/main.js
@@ -1,151 +1,1 @@
-'use strict';
-
-const OFFER_TITLES = [
-  'Private room in guesthouse Soul',
-  'Riverbank apartment',
-  'Villa Sakura',
-  'Chalet Hidden gem',
-];
-
-const OFFER_DESCRIPTIONS = [
-  'Clean and cosy private room in the center of Tokyo. Accommodation has a dining area, a flat-screen TV and a private bathroom with shower and a hairdryer.',
-  'This accommodation is situated on Gyotoku Riverbank in 300 metres from all local attractions. The apartment offers a children\'s playground. Fishing and hiking can be enjoyed nearby.',
-  'In the heart of Tokyo, set within a short distance of Ruins of Takanawa Okido Gate and Shinsei Temple, this property is a renovated Japanese townhouse fitted with air conditioning and free WiFi. Guests can also relax in the shared lounge area. Free private parking is available.',
-  'Suitable location in an area where guests can engage in activities such as hiking and cycling. This air-conditioned property is composed of 2 separate bedrooms, a living room, a fully equipped kitchen with a microwave and fridge, and 1 bathroom.',
-];
-
-const HOUSING_TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-];
-
-const CHECK_TIME = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const FEATURES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-];
-
-const PHOTOS = [
-  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
-  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
-  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
-];
-
-const MIN_LATITUDE = 35.65000;
-const MAX_LATITUDE = 35.70000;
-const MIN_LONGITUDE = 139.70000;
-const MAX_LONGITUDE = 139.80000;
-const LOCATION_FLOAT = 5;
-const SIMILAR_OFFER_COUNT = 10;
-
-// Функция, возвращающая случайное целое число из переданного диапазона включительно.
-const getRandomIntInclusive = (min, max) => {
-  if (min >= max || min < 0 || max < 0) {
-    return null;
-  }
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-};
-
-// Функция, возвращающая случайное число с плавающей точкой из переданного диапазона включительно.
-const getRandomFloatInclusive = (min, max, numbersAfterPoint) => {
-  if (min >= max || min < 0 || max < 0) {
-    return null;
-  }
-  const result = Math.random() * (max - min) + min;
-  return result.toFixed(numbersAfterPoint);
-};
-
-// Функция, возвращающая случайный элемент массива.
-const getRandomArrayElement = (elements) => {
-  const randomArrayIndex = getRandomIntInclusive(0, elements.length - 1);
-  return elements[randomArrayIndex];
-};
-
-// Тасование Фишера — Йетса https://learn.javascript.ru/task/shuffle
-const shuffle = (array) => {
-  for (let i = array.length - 1; i > 0; i--) {
-    let j = Math.floor(Math.random() * (i + 1));
-    [array[i], array[j]] = [array[j], array[i]];
-  }
-};
-
-// Функция, возвращающая массив случайной длины из значений массива (значения исходного массива предварительно тасуются).
-const getRandomArray = (array) => {
-  const randomArrayIndex = getRandomIntInclusive(0, array.length - 1);
-  shuffle(array);
-  return array.slice(randomArrayIndex);
-};
-
-/*
-// Функция, возвращающая массив случайной длины из значений массива (значения исходного массива предварительно тасуются), в.2.
-
-const getRandomArray = (array) => {
-  const randomArrayIndex = getRandomIntInclusive(0, array.length - 1);
-  return  array.sort(() => Math.random() > 0.5 ? 1 : -1).slice(randomArrayIndex);
-};
-
-// Функция, возвращающая массив случайной длины из значений, без перемешивания исходного массива, в.1.
-
-const getRandomArray = (array) => {
-  let newArray = [];
-  array.forEach((element) => {
-    if (getRandomIntInclusive(0, 1)) {
-      return;
-    }
-    newArray.push(element);
-  })
-  return newArray;
-}
-
-// Функция, возвращающая массив случайной длины из значений, без перемешивания исходного массива, в.2.
-
-const getRandomArray = (array) => {
-  const randomArrayIndex = getRandomIntInclusive(0, array.length - 1);
-  return array.slice(randomArrayIndex);
-};
-*/
-
-const getRandomLocation = () => {
-  const x = getRandomFloatInclusive(MIN_LATITUDE, MAX_LATITUDE, 5);
-  const y = getRandomFloatInclusive(MIN_LONGITUDE, MAX_LONGITUDE, 5);
-  return Number(x) + ', ' + Number(y);
-};
-
-const createOffer = () => {
-  return {
-    author: {
-      avatar: `img/avatars/user0${getRandomIntInclusive(1, 8)}.png`,
-    },
-    offer: {
-      title: getRandomArrayElement(OFFER_TITLES),
-      address: getRandomLocation(),
-      price: getRandomIntInclusive(10,2000),
-      type: getRandomArrayElement(HOUSING_TYPES),
-      rooms: getRandomIntInclusive(1,200),
-      guests: getRandomIntInclusive(1,450),
-      checkin: getRandomArrayElement(CHECK_TIME),
-      checkout: getRandomArrayElement(CHECK_TIME),
-      features: getRandomArray(FEATURES),
-      description: getRandomArrayElement(OFFER_DESCRIPTIONS),
-      photos: getRandomArray(PHOTOS),
-    },
-    location: {
-      x: +getRandomFloatInclusive(MIN_LATITUDE, MAX_LATITUDE, LOCATION_FLOAT),
-      y: +getRandomFloatInclusive(MIN_LONGITUDE, MAX_LONGITUDE, LOCATION_FLOAT),
-    },
-  };
-};
-
-const createSimilarOffers = new Array(SIMILAR_OFFER_COUNT).fill(null).map(() => createOffer());
-createSimilarOffers();
+import './data.js';
